+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ and by implication, [Semantic Versioning](http://semver.org/).
 - Updated get_checklist to return duration in minutes.
 - Updated get_checklist to return distance in metres.
 - Updated get_checklist to return area in hectares.
+- Updated get_checklist to include observation comments.
 
 [0.2.0] - 2024-10-23
 --------------------